update:MVCmodel

+++ app/application.go
@@ -97,7 +97,7 @@ func MemberAdd(db *gorm.DB) echo.HandlerFunc {
 			return c.JSON(404, em)
 		}
 		db.Create(&member)
-		name := c.QueryParam("project")
+		name := c.QueryParam("team")
 		if strings.Contains(name, "(") && strings.Contains(name, ")") {
 			new_pro := strings.Split(name[1:len(name)-1], ",")
 			for _, v := range new_pro {
@@ -156,7 +156,7 @@ func MemberUpdate(db *gorm.DB) echo.HandlerFunc {
 		grade := Grade{Name: c.QueryParam("new_grade")}
 		db.First(&grade, "name = ?", grade.Name)
 		new_data := Member{id, c.QueryParam("new_name"), grade.Id, c.QueryParam("new_mail_address")}
-		new_project := Project{0, c.QueryParam("new_project")}
+		new_project := Project{0, c.QueryParam("new_team")}
 		old_data := db.Model(&member)
 		if new_data.Name != "" {
 			old_data.Update("name", new_data.Name)
@@ -170,34 +170,39 @@ func MemberUpdate(db *gorm.DB) echo.HandlerFunc {
 		if new_project.Name != "" {
 			name := new_project.Name
 			member_project := []Member_Project{}
+			project := Project{}
 			if strings.Contains(name, "(") && strings.Contains(name, ")") {
 				new_pro := strings.Split(name[1:len(name)-1], ",")
 				for _, v := range new_pro {
-					project := Project{}
+					// project := Project{}
 					db.First(&project, "name = ?", v)
 					if ErrorCheck(project.Id) != true {
-						em := Error{Massege: "not exit grade"}
+						em := Error{Massege: "not exit project"}
 						return c.JSON(404, em)
 					}
 					db.Create(&Member_Project{Member_Id: member.Id, Project_Id: project.Id})
 				}
 			} else {
-				project := Project{}
+				//project := Project{}
 				db.First(&project, "name = ?", name)
 				fmt.Println(project)
 				if ErrorCheck(project.Id) != true {
-					em := Error{Massege: "not exit grade"}
+					em := Error{Massege: "not exit project"}
 					return c.JSON(404, em)
 				}
 				db.Create(&Member_Project{Member_Id: member.Id, Project_Id: project.Id})
 			}
+
 			db.Find(&member_project, "member_id = ?", member.Id)
 			for _, v := range member_project {
 				if ErrorCheck(v.Member_Id) != true {
 					em := Error{Massege: "not exit Member"}
 					return c.JSON(404, em)
 				}
-				db.Delete(&v)
+				if v.Project_Id != project.Id {
+					fmt.Println(v)
+					db.Delete(&v)
+				}
 			}
 		}
 		return c.JSON(http.StatusOK, Error{Massege: "OK"})
@@ -280,5 +285,5 @@ func main() {
 	e.Get("/api/show/address_count", ShowAddressCount(db))
 
 	// サーバー構築 ポート8000
-	e.Run(standard.New(":1323"))
+	e.Run(standard.New(":50000"))
 }
